refactor(test): split triage tests.rs god-file into tests/ subdir
